+++ src/silk/enc_API.rs
@@ -536,7 +536,7 @@ pub unsafe fn silk_Encode(
                             (*psEnc).state_Fxx[n as usize].sCmn.indices_LBRR[i as usize]
                                 .quantOffsetType as i32,
                             &mut (*psEnc).state_Fxx[n as usize].sCmn.pulses_LBRR[i as usize],
-                            (*psEnc).state_Fxx[n as usize].sCmn.frame_length,
+                            (*psEnc).state_Fxx[n as usize].sCmn.frame_length as usize,
                         );
                     }
                     n += 1;
