+++ src/silk/float/encode_frame_FLP.rs
@@ -248,7 +248,7 @@ pub unsafe fn silk_encode_frame_FLP(
                     (*psEnc).sCmn.indices.signalType as i32,
                     (*psEnc).sCmn.indices.quantOffsetType as i32,
                     &mut (*psEnc).sCmn.pulses,
-                    (*psEnc).sCmn.frame_length,
+                    (*psEnc).sCmn.frame_length as usize,
                 );
                 nBits = ec_tell(psRangeEnc);
                 if iter == maxIter && found_lower == 0 && nBits > maxBits {
@@ -281,7 +281,7 @@ pub unsafe fn silk_encode_frame_FLP(
                         (*psEnc).sCmn.indices.signalType as i32,
                         (*psEnc).sCmn.indices.quantOffsetType as i32,
                         &mut (*psEnc).sCmn.pulses,
-                        (*psEnc).sCmn.frame_length,
+                        (*psEnc).sCmn.frame_length as usize,
                     );
                     nBits = ec_tell(psRangeEnc);
                 }
